requirements.txt
Flask
PyPDF2
python-docx
pdf2image
pytesseract
Pillow
